Fix some grammar, add a footnote...
Maybe I should expand that section more and maybe add a section in ch5
about how to use fp and/or pass arguments on the stack if necessary that
I can refer to as "we'll cover making local variables on the stack in
ch5".

It just...seems unecessary and beyond the scope of the book since it's
almost never used.